+++ src/main/java/finalmission/woowabowling/auth/service/response/LoginMember.java
@@ -1,7 +1,6 @@
-package finalmission.woowabowling.member;
+package finalmission.woowabowling.auth.service.response;
 
 public record LoginMember(
-
         Long id,
         String name,
         String email) {
