+++ src/main/java/finalmission/woowabowling/lane/controller/request/LaneRegisterRequest.java
@@ -1,4 +1,4 @@
-package finalmission.woowabowling.lane;
+package finalmission.woowabowling.lane.controller.request;
 
 public record LaneRegisterRequest(
         int number,
