--- src/main/java/finalmission/woowabowling/member/LoginRequest.java
@@ -1,13 +1,0 @@
-package finalmission.woowabowling.member;
-
-import finalmission.woowabowling.member.domain.Member;
-
-public record LoginRequest(
-
-        String email,
-        String password
-) {
-    public Member toMember(final String name) {
-        return Member.from(name, email, password);
-    }
-}
