+++ src/main/java/finalmission/woowabowling/member/controller/request/LoginRequest.java
@@ -1,0 +1,7 @@
+package finalmission.woowabowling.member.controller.request;
+
+public record LoginRequest(
+        String email,
+        String password
+) {
+}
