+++ src/main/java/finalmission/woowabowling/reservatoin/service/ReservationService.java
@@ -4,11 +4,11 @@ import finalmission.woowabowling.lane.domain.Lane;
 import finalmission.woowabowling.lane.domain.LaneRepository;
 import finalmission.woowabowling.member.domain.Member;
 import finalmission.woowabowling.member.domain.MemberRepository;
-import finalmission.woowabowling.reservatoin.ReservationRegisterRequest;
-import finalmission.woowabowling.reservatoin.ReservationRegisterResponse;
-import finalmission.woowabowling.reservatoin.ReservationUpdateRequest;
+import finalmission.woowabowling.reservatoin.controller.request.ReservationRegisterRequest;
+import finalmission.woowabowling.reservatoin.controller.request.ReservationUpdateRequest;
 import finalmission.woowabowling.reservatoin.domain.Reservation;
 import finalmission.woowabowling.reservatoin.domain.ReservationRepository;
+import finalmission.woowabowling.reservatoin.service.response.ReservationRegisterResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
