+++ src/test/java/finalmission/woowabowling/reservatoin/controller/MemberReservationControllerTest.java
@@ -10,11 +10,11 @@ import finalmission.woowabowling.member.domain.Member;
 import finalmission.woowabowling.member.domain.MemberRepository;
 import finalmission.woowabowling.pattern.domain.Pattern;
 import finalmission.woowabowling.pattern.domain.PatternRepository;
-import finalmission.woowabowling.reservatoin.ReservationRegisterRequest;
-import finalmission.woowabowling.reservatoin.ReservationRegisterResponse;
-import finalmission.woowabowling.reservatoin.ReservationUpdateRequest;
+import finalmission.woowabowling.reservatoin.controller.request.ReservationRegisterRequest;
+import finalmission.woowabowling.reservatoin.controller.request.ReservationUpdateRequest;
 import finalmission.woowabowling.reservatoin.domain.Reservation;
 import finalmission.woowabowling.reservatoin.domain.ReservationRepository;
+import finalmission.woowabowling.reservatoin.service.response.ReservationRegisterResponse;
 import io.restassured.RestAssured;
 import io.restassured.common.mapper.TypeRef;
 import io.restassured.http.ContentType;
