+++ src/test/java/finalmission/woowabowling/reservatoin/controller/ReservationControllerTest.java
@@ -8,9 +8,9 @@ import finalmission.woowabowling.member.domain.Member;
 import finalmission.woowabowling.member.domain.MemberRepository;
 import finalmission.woowabowling.pattern.domain.Pattern;
 import finalmission.woowabowling.pattern.domain.PatternRepository;
-import finalmission.woowabowling.reservatoin.ReservationRegisterResponse;
 import finalmission.woowabowling.reservatoin.domain.Reservation;
 import finalmission.woowabowling.reservatoin.domain.ReservationRepository;
+import finalmission.woowabowling.reservatoin.service.response.ReservationRegisterResponse;
 import io.restassured.RestAssured;
 import io.restassured.common.mapper.TypeRef;
 import io.restassured.http.ContentType;
